move example utils
